add CommandRegistry

+++ src/main/java/disneyprincess/commands/DeleteCommand.java
@@ -1,5 +1,6 @@
 package disneyprincess.commands;
 
+import disneyprincess.model.Princess;
 import disneyprincess.repository.PrincessRepository;
 
 public class DeleteCommand implements Command {
@@ -7,12 +8,24 @@ public class DeleteCommand implements Command {
 
   public DeleteCommand(PrincessRepository repository) {
     this.repository = repository;
-
   }
 
   @Override
   public void execute(String[] args) {
+    if (args.length < 1) {
+      System.out.println("You must specify the princess ID to delete");
+      return;
+    }
+
     int deleteId = Integer.parseInt(args[0]);
+    Princess princess = repository.get(deleteId);
+
+    if (princess == null) {
+      System.out.println("There are no princesses with this ID");
+      return;
+    }
+
     repository.delete(deleteId);
+    System.out.println("Princess deleted");
   }
 }
